Refs #4703 Move FastMiddleware inside fast folder.
